Solution to Digit Sequence Mountain Peak Finder

+++ Digit Sequence Mountain Peak Finder@@67de00956ac3fec9afb90e11/code.c
@@ -20,7 +20,7 @@ int mountainPeak(int n)
 
     if(l==i)
     {
-        return s[l];
+        return (int)s[l];
     }
 
     //printf("%d %d",l,i);
